# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.MenuMeta:
    properties:
      auth:
        type: boolean
      caption:
        type: string
      component:
        type: string
      disabled:
        type: boolean
      external_link:
        type: string
      hidden:
        type: boolean
      icon:
        type: string
      info:
        type: string
    type: object
  entity.MenuStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - MenuStatusActive
    - MenuStatusInactive
  entity.MenuType:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      MenuTypeButton: 按钮
      MenuTypeCatalogue: 目录
      MenuTypeGroup: 分组
      MenuTypeMenu: 菜单
    x-enum-descriptions:
    - 分组
    - 目录
    - 菜单
    - 按钮
    x-enum-varnames:
    - MenuTypeGroup
    - MenuTypeCatalogue
    - MenuTypeMenu
    - MenuTypeButton
...
  model.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  model.AssignIDsRequest:
    properties:
      menu_ids:
        example:
        - "1"
        - "2"
        items:
          type: string
        type: array
      permission_ids:
        example:
        - "1"
        - "2"
        items:
          type: string
        type: array
      role_ids:
        example:
        - "1"
        - "2"
        items:
          type: string
        type: array
    type: object
...
  service.CreateMenuRequest:
    properties:
      code:
        maxLength: 100
        type: string
      description:
        maxLength: 500
        type: string
      meta:
        $ref: '#/definitions/entity.MenuMeta'
      name:
        maxLength: 100
        type: string
      order:
        type: integer
      parent_id:
        type: string
      path:
        maxLength: 200
        type: string
      status:
        $ref: '#/definitions/entity.MenuStatus'
      type:
        $ref: '#/definitions/entity.MenuType'
    required:
    - code
    - name
    - type
    type: object
...
  service.UpdateMenuRequest:
    properties:
      code:
        maxLength: 100
        type: string
      description:
        maxLength: 500
        type: string
      id:
        type: string
      meta:
        $ref: '#/definitions/entity.MenuMeta'
      name:
        maxLength: 100
        type: string
      order:
        type: integer
      parent_id:
        type: string
      path:
        maxLength: 200
        type: string
      status:
        $ref: '#/definitions/entity.MenuStatus'
      type:
        $ref: '#/definitions/entity.MenuType'
    required:
    - code
    - id
    - name
    - type
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/coze-dev/alice
  description: Alice 企业级后端 API 文档。
  title: Alice API
  version: "1.0"
paths:
...
  /menus:
    get:
      description: 获取所有菜单列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 菜单列表
      tags:
      - Menus
    post:
      consumes:
      - application/json
      description: 创建一个新的菜单
      parameters:
      - description: 创建菜单请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateMenuRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 创建菜单
      tags:
      - Menus
  /menus/{id}:
    delete:
      description: 删除指定菜单
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除菜单
      tags:
      - Menus
    get:
      description: 根据ID获取菜单详情
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取菜单详情
      tags:
      - Menus
    put:
      consumes:
      - application/json
      description: 更新指定菜单
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新菜单请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 更新菜单
      tags:
      - Menus
  /menus/tree:
    get:
      description: 获取完整菜单树
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 菜单树
      tags:
      - Menus
...
  /roles/{id}/menus:
    delete:
      consumes:
      - application/json
      description: 从角色移除一组菜单
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      - description: 菜单ID集合 (menu_ids)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AssignIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 移除角色菜单
      tags:
      - RoleMenus
    get:
      description: 获取指定角色的菜单列表
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取角色菜单
      tags:
      - RoleMenus
    post:
      consumes:
      - application/json
      description: 为指定角色批量分配菜单
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      - description: 菜单ID集合 (menu_ids)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AssignIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 为角色分配菜单
      tags:
      - RoleMenus
...
  /users/{user_id}/menus:
    get:
      description: 获取指定用户的菜单列表
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取用户菜单
      tags:
      - UserMenus
  /users/{user_id}/menus/tree:
    get:
      description: 获取指定用户的菜单树
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 用户菜单树
      tags:
      - UserMenus
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: '请输入形如: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"